Se Agrego validaciones de cliente, proveedor y vista si es admin, usuario o consultor

+++ view/cliente/indexClienteTemplate.html.php
@@ -2,6 +2,7 @@
 <?php use mvc\routing\routingClass as routing ?>
 <?php use mvc\i18n\i18nClass as i18n ?>
 <?php use mvc\view\viewClass as view?>
+<?php use mvc\session\sessionClass as session?>
 <?php $nom = clienteTableClass::NOMBRE ?>
 <?php $apellido = clienteTableClass::APELLIDO ?>
 <?php $id = clienteTableClass::ID ?>
@@ -22,8 +23,10 @@
     <article id='derecha'>
        <h1><?php echo i18n::__('cliente') ?></h1>
       <ul>
-     <a class="btn  btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'insertCliente') ?>"><img class="img-responsive"  id="imgnuevo" src="" alt=" "><?php echo i18n::__('nuevo') ?></a> 
+     <?php if (session::getInstance()->hasCredential('admin')):?>
+        <a class="btn  btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'insertCliente') ?>"><img class="img-responsive"  id="imgnuevo" src="" alt=" "><?php echo i18n::__('nuevo') ?></a> 
         <a href="javascript:eliminarMasivo()" class="btn  btn btn-xs" id="btnDeleteMasivo"><img class="img-responsive"  id="imgmasivo" src="" alt=" "><?php echo i18n::__('eliminar en masa') ?></a> 
+        <?php endif?>
         <a type="button" class="btn  btn btn-xs" data-toggle="modal" data-target="#myModalFilters"><img class="img-responsive"  id="imgfiltros" src="" alt=" "><?php echo i18n::__('filtros') ?></a>
         <a href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'indexCliente') ?>" class="btn  btn-xs" ><img class="img-responsive"  id="imgelifiltro" src="" alt=" "><?php echo i18n::__('eFiltros') ?></a>
         <a type="button" class="btn  btn-xs" data-toggle="modal" data-target="#myModalReport" ><img class="img-responsive"  id="imgreporte" src="" alt=" "><?php echo i18n::__('informe') ?></a>          
@@ -130,7 +133,8 @@
       
       <form class="form-signin" id="frmDeleteAll" action="<?php echo routing::getInstance()->getUrlWeb('cliente', 'deleteSelectCliente') ?>" method="POST">        
         <?php view::includeHandlerMessage()?>       
-        <table class="table table-bordered table-responsive">
+        <br>
+        <table id="tabla" class="table table-bordered table-responsive">
           <tr>
           <thead>
           <th>
@@ -160,9 +164,11 @@
                   <?php echo $key->$apellido ?>
                 </td>
                 <th>
-                  <a class="btn btn-warning btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'verCliente', array(clienteTableClass::ID => $key->$id)) ?>" > <?php echo i18n::__('ver') ?></a> -
-                  <a class="btn btn-primary btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'editCliente', array(clienteTableClass::ID => $key->$id)) ?>"> <?php echo i18n::__('modificar') ?> </a> -
+                  <a class="btn btn-warning btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'verCliente', array(clienteTableClass::ID => $key->$id)) ?>" > <?php echo i18n::__('ver') ?></a>
+                  <?php if (session::getInstance()->hasCredential('admin')):?>
+                  <a class="btn btn-primary btn-xs" href="<?php echo routing::getInstance()->getUrlWeb('cliente', 'editCliente', array(clienteTableClass::ID => $key->$id)) ?>"> <?php echo i18n::__('modificar') ?> </a>
                   <a data-toggle="modal" data-target="#myModalDelete<?php echo $key->$id ?>" class="btn btn-danger btn-xs"> <?php echo i18n::__('eliminar') ?></a>
+                <?php endif?>
                 </th>
        </tr>
   <div class="modal fade" id="myModalDelete<?php echo $key->$id ?>" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
